feat(middleware): add internet online check before accessing Google Services
